The liquibase client is moved to the kapua-commons package.
