+++ commons/src/main/java/org/eclipse/kapua/commons/liquibase/KapuaLiquibaseClient.java
@@ -10,7 +10,7 @@
  *     Red Hat Inc - initial API and implementation
  *     Eurotech
  *******************************************************************************/
-package org.eclipse.kapua.service.liquibase;
+package org.eclipse.kapua.commons.liquibase;
 
 import liquibase.Liquibase;
 import liquibase.database.Database;
