+++ service/account/internal/pom.xml
@@ -43,10 +43,6 @@
             <groupId>org.eclipse.kapua</groupId>
             <artifactId>kapua-guice</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.eclipse.kapua</groupId>
-            <artifactId>kapua-liquibase</artifactId>
-        </dependency>
 
         <!-- Test dependencies -->
         <dependency>
