+++ service/job/internal/pom.xml
@@ -52,10 +52,6 @@
             <groupId>org.eclipse.kapua</groupId>
             <artifactId>kapua-guice</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.eclipse.kapua</groupId>
-            <artifactId>kapua-liquibase</artifactId>
-        </dependency>
 
         <!-- Test dependencies -->
         <dependency>
