+++ service/pom.xml
@@ -25,8 +25,6 @@
     <artifactId>kapua-service</artifactId>
 
     <modules>
-        <module>liquibase</module>
-
         <module>api</module>
 
         <module>account</module>
