+++ service/security/shiro/src/test/java/org/eclipse/kapua/service/authorization/shiro/AbstractAuthorizationServiceTest.java
@@ -27,7 +27,7 @@ import org.eclipse.kapua.commons.setting.system.SystemSettingKey;
 import org.eclipse.kapua.locator.KapuaLocator;
 import org.eclipse.kapua.model.id.KapuaId;
 import org.eclipse.kapua.service.authentication.shiro.AuthenticationEntityManagerFactory;
-import org.eclipse.kapua.service.liquibase.KapuaLiquibaseClient;
+import org.eclipse.kapua.commons.liquibase.KapuaLiquibaseClient;
 import org.eclipse.kapua.test.junit.JUnitTests;
 import org.junit.Assert;
 import org.junit.experimental.categories.Category;
