+++ service/user/internal/pom.xml
@@ -44,10 +44,6 @@
             <groupId>org.eclipse.kapua</groupId>
             <artifactId>kapua-guice</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.eclipse.kapua</groupId>
-            <artifactId>kapua-liquibase</artifactId>
-        </dependency>
 
         <!-- Test dependencies -->
         <dependency>
